# fix(examples): use anonymous access when reading example data from GCS