Fixes #8093.