@jcrist 2024-01-25
This also seems to the performance problems I was having loading example data.

Before this change:

```python
In [2]: %time t = ibis.examples.imdb_title_basics.fetch()
CPU times: user 891 ms, sys: 673 ms, total: 1.56 s
Wall time: 22.3 s
```

With this change:

```python
In [2]: %time t = ibis.examples.imdb_title_basics.fetch()
CPU times: user 893 ms, sys: 650 ms, total: 1.54 s
Wall time: 1.06 s
```

Looking deeper at the profiling information, most of the time in the first instance is in setting up the board (probably something is weird with my local credentials, idk).